requirements.txt
alabaster==0.7.12
alembic==1.0.5
attrs==18.2.0
autoenv==1.0.0
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cliff==2.14.0
cmd2==0.9.6
colorama==0.4.1
docutils==0.14
dominate==2.3.5
filelock==3.0.10
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gnupg==2.3.1
idna==2.8
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.15.4
packaging==18.0
pbr==5.1.1
Pillow==5.3.0
pluggy==0.8.0
prettytable==0.7.2
psutil==5.4.8
psycopg2-binary==2.7.6.1
py==1.7.0
Pygments==2.3.1
pyparsing==2.3.0
pyperclip==1.7.0
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
python-secrets==18.11.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.15
stevedore==1.30.0
toml==0.10.0
tox==3.6.0
urllib3==1.24.1
virtualenv==16.1.0
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
xkcdpass==1.16.5
yamlreader==3.0.4
